change int by size_t
git-svn-id: file:///home/roberto/Git/svn@6536 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 83af4414dea571fce9ea7cb614a73836181dd345

+++ trunk/xmipp/libraries/data/rwIMAGIC.h
@@ -306,7 +306,7 @@ int  writeIMAGIC(int img_select=-1, int mode=WRITE_OVERWRITE)
         REPORT_ERROR(ERR_TYPE_INCORRECT,"ERROR write IMAGIC image: invalid typeid(T)");
 
     size_t datasize, datasize_n;
-    datasize_n = Xdim*Ydim*Zdim;
+    datasize_n = (size_t)Xdim*Ydim*Zdim;
     if (isComplexT())
         datasize = datasize_n * gettypesize(ComplexFloat);
     else
